fix: Move the remote joiner class to the modules SDK package
