+++ packages/core/modules-sdk/package.json
@@ -38,7 +38,6 @@
   },
   "dependencies": {
     "@medusajs/deps": "2.17.1",
-    "@medusajs/orchestration": "2.17.1",
     "@medusajs/utils": "2.17.1"
   }
 }
