+++ packages/core/modules-sdk/src/__fixtures__/joiner/data.ts
