+++ packages/core/modules-sdk/src/__tests__/joiner/remote-joiner.spec.ts
@@ -5,7 +5,7 @@ import {
 } from "@medusajs/types"
 import { lowerCaseFirst, toPascalCase } from "@medusajs/utils"
 import { serviceConfigs, serviceMock } from "../../__mocks__/joiner/mock_data"
-import { RemoteJoiner } from "./../../joiner"
+import { RemoteJoiner } from "../../joiner"
 
 const container = {
   resolve: (serviceName) => {
