+++ packages/core/modules-sdk/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./definitions"
+export * from "./joiner"
 export * from "./link"
 export * from "./loaders"
 export * from "./medusa-app"
