+++ packages/core/orchestration/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./joiner"
 export * from "./transaction"
 export * from "./workflow"
